Refactoring: Extract TerminationAnalysis to its own file
